refactor: Update Raffle contract to include entrance fee validation
The Raffle contract has been updated to include entrance fee validation in the enterRaffle function. This ensures that the user must send the correct amount of entrance fee to participate in the raffle. The validation is done by checking if the sent value is less than the entrance fee, and reverting the transaction if it is. This improves the functionality and security of the contract.