refactor: Add getLastTimeStamp and getRecentWinner functions to Raffle contract
This commit adds two new functions, getLastTimeStamp and getRecentWinner, to the Raffle contract in the foundry-smart-lottery project. The getLastTimeStamp function returns the last timestamp recorded in the contract, while the getRecentWinner function returns the address of the most recent winner. These functions provide additional information about the state of the Raffle contract and enhance the functionality of the smart lottery project.