refactor: Update DeployRaffle contract to return Raffle and HelperConfig in run() function
The DeployRaffle contract in the foundry-smart-lottery project has been updated to return the Raffle and HelperConfig instances in the run() function. This change allows the contract to deploy the Raffle contract and HelperConfig contract simultaneously and return their instances. By returning these instances, other contracts and scripts can easily access and interact with them. This enhancement improves the functionality and usability of the smart lottery project.